requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
assemblyai==0.19.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
cache==1.0.3
cachelib==0.9.0
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
distro==1.8.0
dnspython==2.4.2
docstring-parser==0.15
dominate==2.8.0
email-validator==2.1.0.post1
emoji==2.8.0
entrypoints==0.4
faiss-cpu==1.7.4
Faker==19.13.0
favicon==0.7.0
filelock==3.13.1
filetype==1.2.0
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Caching==2.1.0
Flask-Login==0.6.2
Flask-Mail==0.9.0
Flask-Migrate==4.0.5
flask-ngrok==0.0.25
Flask-RESTful==0.3.10
Flask-Share==0.1.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.4
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.1
httpx==0.25.1
huggingface==0.0.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
instructor==0.3.5
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.339
langdetect==1.0.9
langsmith==0.0.66
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.1
openai==1.3.3
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
psycopg2==2.9.9
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-iso639==2023.6.15
python-magic==0.4.27
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
razorpay==1.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-rust==1.7.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
st-annotated-text==4.0.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.30
urllib3==2.0.6
validators==0.22.0
visitor==0.1.3
watchdog==3.0.0
websockets==12.0
Werkzeug==2.3.0
wsproto==1.2.0
yarl==1.9.2
zerobouncesdk==1.1.0
zipp==3.17.0


# pip install huggingface
# pip install transformers
# pip install InstructorEmbedding
# pip install torch
# pip install sentence_transformers